+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 - Fix entry preview error on keyword checklist (Fixes #224)
+- Fix SEO field error when changing entry types (Fixes #215)
 
 ## 3.5.4 - 2019-01-28