+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix entire image meta data being saved with SEO field (Fixes #142)
 - Fix sitemap error when SEO field doesn't have any advanced data (Fixes #96)
 - Fix keyword checklist reading title length incorrectly (Fixes #199)
+- Fix social meta outputting encoded characters (Fixes #198)
 
 ## 3.5.4 - 2019-01-28