+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Improved
 - Improve alt text judgment in keyword checklist
 - Check for keyword in first paragraph of body, main, and article tags
+- Improve page meta field instructions in SEO settings
 
 ## 3.5.4 - 2019-01-28